requirements.txt
## The following requirements were added by pip freeze:
Django==2.1.7
djangorestframework==3.9.1
pytz==2018.9
selenium==3.141.0
urllib3==1.24.1
sqlparse==0.2.4
